recommendations service is functional with postgres
I had to comment out some image receipt code in the carousel of component did update in order to get images to render, but it was messy.
I consider it good enough because my service is not the source of those photos anyways.  The old code is only commented out so I can restore it when the time comes to get the information from the image service endpoint